alert-apagar

+++ listartarefas.php
@@ -7,6 +7,7 @@ if(!isset($_SESSION['usuario_id'])){
     exit();
 }
 ?>
+
 <!DOCTYPE html>
 <html lang="pt-br">
 <head>
@@ -27,6 +28,7 @@ if(!isset($_SESSION['usuario_id'])){
             <img src="img/logo.png" alt="Logo" class="logo">
             <span class="nome-empresa">W E E K</span>
         </div>
+        
         <a href="home_nova.php" class="nav-home">Home</a>
     </header>
     <nav class="segunda-navbar">
@@ -35,51 +37,51 @@ if(!isset($_SESSION['usuario_id'])){
         </a>
         <span class="nav-calendario"><a class="nav-calendario" href="calendariooficial.php">Calendário</a></span>
     </nav>
+
     <?php
-include 'conexao.php';
+    include 'conexao.php';
 
-$sql = "SELECT * FROM tarefas INNER JOIN funcionario ON tarefas.fk_cliente_id = funcionario_id  ORDER BY data_tarefa ASC";
-$resultado = $conn->query($sql);
+    $sql = "SELECT * FROM tarefas INNER JOIN funcionario ON tarefas.fk_cliente_id = funcionario_id  ORDER BY data_tarefa ASC";
+    $resultado = $conn->query($sql);
 
-echo "<br><h1>Lista de Tarefas</h1>";
+    echo "<br><h1>Lista de Tarefas</h1>";
 
+    if ($resultado->num_rows > 0) {
+        echo "<div class='table-container'>";
+        echo "<table border='2'>";
+        echo "<tr><th>Nome do funcionário</th><th>Nome da Tarefa</th><th>Data em que será realizada</th><th>Status</th><th>Funcionário que realizará</th><th>Ações</th></tr>";
 
-if ($resultado->num_rows > 0) {
-    echo "<div class='table-container'>";
-    echo "<table border='2'>";
-    echo "<tr><th>Nome do funcionário</th><th>Nome da Tarefa</th><th>Data em que será realizada</th><th>Status</th><th>Funcionário que realizara</th><th>Ações</th></tr>";
+        while ($row = $resultado->fetch_assoc()) {
+            echo "<tr>";
+            echo "<td>" . $row['funcionario_nome'] . "</td>";
+            echo "<td>" . $row['descricao_tarefa'] . "</td>";
+            echo "<td>" . $row['data_tarefa'] . "</td>";
+            echo "<td>" . $row['status_tarefa'] . "</td>";
+            echo "<td>" . $row['funcionario_nome'] . "</td>";
+            echo "<td>";
+            echo "<a href='editar.php?id=" . $row['id_tarefa'] . "'><span class='material-icons'>edit</span></a> | ";
+            echo "<a href='excluir.php?id=" . $row['id_tarefa'] . "' onclick='return confirmarExclusao();'><span class='material-icons'>delete</span></a>";
+            echo "</td>";
+            echo "</tr>";
+        }
 
-    while ($row = $resultado->fetch_assoc()) {
-        echo "<tr>";
-        echo "<td>" . $row['funcionario_nome'] . "</td>";
-        echo "<td>" . $row['descricao_tarefa'] . "</td>";
-        echo "<td>" . $row['data_tarefa'] . "</td>";
-        echo "<td>" . $row['status_tarefa'] . "</td>";
-        echo "<td>" . $row['funcionario_nome'] . "</td>";
-        echo "<td>";
-        echo "<a href='editar.php?id=" . $row['id_tarefa'] . "'><span class='material-icons'>edit</span></a> | ";
-        echo "<a href='excluir.php?id=" . $row['id_tarefa'] . "'><span class='material-icons'>delete</span></a>";
-
-        echo "</td>";
-        echo "</tr>";
+        echo "</table>";
+        echo "</div>"; // Fecha a div table-container
+    } else {
+        echo "Nenhuma tarefa listada.";
     }
 
-    echo "</table>";
-    echo "</div>"; // Close the table-container div
-} else {
-    echo "Nenhuma tarefa listada.";
-}
+    $conn->close();
 
-$conn->close();
+    ?>
 
-?>
-<br>
-<div class="btn-volta-container">
-    <a href="home_nova.php" class="btn-volta">Voltar</a>
-</div>
-<br>
+    <br>
+    <div class="btn-volta-container">
+        <a href="home_nova.php" class="btn-volta">Voltar</a>
+    </div>
+    <br>
 
-<footer>
+    <footer>
     <div class="conteudogeral">
         <div class="conteudo1">
             <h4>Sobre nós</h4><br>
@@ -110,42 +112,14 @@ $conn->close();
             </div>
         </div>
     </div>
-    </div>
-</footer>
+    </div>    </footer>
+
+
+    <script>
+    function confirmarExclusao() {
+        return confirm("Tem certeza de que deseja excluir esta tarefa?");
+    }
+    </script>
+
 </body>
 </html>
-
-
-
-
-
-<!-- // include 'conexao.php';
-
-// $sql = "SELECT * FROM paciente";
-// $resultado = $conexao->query($sql);
-
-// if ($resultado->num_rows > 0) {
-//     while ($row = $resultado->fetch_assoc()) {
-
-//     }
-// } else {
-//     echo "Nenhum paciente cadastrado.";
-// }
-
-// echo "<table>";
-// echo "<tr><th>ID</th><th>Nome</th><th>Idade</th><th></tr>";
-// while ($row = $resultado->fetch_assoc()) {
-//     echo "<tr>";
-//     echo "td" . $row['id_paciente'] . "</td>";
-//     echo "td" . $row['nome_paciente'] . "</td>";
-//     echo "td" . $row['cpf_paciente'] . "</td>";
-//     echo "td" . $row['convenio_paciente'] . "</td>";
-// }
-
-// echo "</table>";
-
-// if ($resultado->num_rows === 0) {
-//     echo "Nenhum paciente cadastrado";
-// }
-
-// $conexao->close(); -->
